+++ src/templates/hooks/useUixTemplate.ts
@@ -9,7 +9,7 @@ import {  validateDraftSchema, DraftErrorTree, AnyNodeDefinitionMap, SubgraphDef
 import { extractSubgraph, mergeSubgraph } from "./functionModule"
 import { ZodObject, ZodTypeAny, z, AnyZodObject, ZodIssue } from "zod"
 import { useImmer } from "@thinairthings/use-immer"
-import { useEffect, useRef } from "react"
+import { useCallback, useEffect, useRef } from "react"
 import { produce, WritableDraft } from "immer"
 import _ from "lodash"
 
@@ -50,7 +50,7 @@ export const useUix = <
     ) => Data
 }) => {
     const queryClient = useQueryClient()
-    const queryResult = useQuery({
+    const {data, isPending} = useQuery({
         queryKey: rootNodeIndex ? [{
             rootNodeIndex: {
                 nodeType: rootNodeIndex.nodeType,
@@ -88,10 +88,15 @@ export const useUix = <
                 })
             }
             addNodeToCache(subgraph)
+            initialDraftRef.current = ((initializeDraft && subgraph) 
+                ? initializeDraft(result.data, (initializedDraft) => initializedDraft) 
+                : subgraph
+            ) as any
+            updateDraft(initialDraftRef.current)
             return result.data 
         } : skipToken
     })
-    const subgraph = queryResult.data
+    const subgraph = data
     const initialDraftRef = useRef(((initializeDraft && subgraph) 
         ? initializeDraft(subgraph, (initializedDraft) => initializedDraft) 
         : subgraph
@@ -99,10 +104,9 @@ export const useUix = <
     const [draft, updateDraft] = useImmer(initialDraftRef.current)
     const [draftErrors, setDraftErrors] = useImmer({} as DraftErrorTree<Data>)
     const mutation = useMutation({
-        mutationFn: async () => {
+        mutationFn: async (data:Data) => {
             if (!draft || !subgraph) return
-            if (_.isEqual(draft, initialDraftRef.current)) return
-            await mergeSubgraph(draft as any)
+            await mergeSubgraph(data as any)
             return null as any
         },
         onMutate: async () => {
@@ -176,30 +180,33 @@ export const useUix = <
         }
     })
     useEffect(() => {
-        if (!subgraph || mutation.isPending) return
-        updateDraft(((initializeDraft && subgraph) 
+        if (!subgraph || mutation.isPending || _.isEqual(initialDraftRef.current, subgraph)) return
+        initialDraftRef.current = ((initializeDraft && subgraph) 
             ? initializeDraft(subgraph, (initializedDraft) => initializedDraft) 
             : subgraph
-        ) as any)
+        ) as any
+        updateDraft(initialDraftRef.current as any)
     }, [subgraph])
     return {
-        ...queryResult,
+        data,
+        isPending,
         draft,
+        draftDidChange: !_.isEqual(draft, initialDraftRef.current),
         draftErrors,
-        resetDraft: () => updateDraft(draft => {
+        resetDraft: useCallback(() => updateDraft(draft => {
             if (!draft) return
             Object.assign(draft, initialDraftRef.current)
-        }),
-        updateDraft: (updater: (callbackDraft: WritableDraft<NonNullable<typeof draft>>) => void) => {
+        }), [draft]),
+        updateDraft: useCallback((updater: (callbackDraft: WritableDraft<NonNullable<typeof draft>>) => void) => {
             if (!draft) return
             updateDraft(draft => {
                 updater(draft as any)
             })
-        },
+        }, [draft]),
         isCommitPending: mutation.isPending,
         isCommitSuccessful: mutation.isSuccess,
         isCommitError: mutation.isError,
-        commitDraft: (options?: Parameters<typeof mutation['mutate']>[1]) => mutation.mutate(undefined, options)
+        commit: (data: Data, options?: Parameters<typeof mutation['mutate']>[1]) => mutation.mutate(data, options)
     }
 }