+++ src/types/index.ts
@@ -2,5 +2,6 @@
 
 
 export * from './ExtractOutputTree'
+export * from './MergeOutputTree'
 export * from './DraftErrorTree'
 export * from './RelationshipKey'