+++ tests/hb.extract.test.ts
@@ -1,7 +1,7 @@
 
 
 import { test } from 'vitest'
-import { extractSubgraph, extractSubgraphv2, mergeSubgraph } from './uix/generated/functionModule'
+import { extractSubgraph } from './uix/generated/functionModule'
 import { throwTestError } from './utils/throwTestError'
 import { BELONGS_TO_Company_Relationship, JobNode, nodeDefinitionMap, UserNode } from './uix/generated/staticObjects'
 
@@ -9,14 +9,17 @@ test('Integration test', async () => {
     // Create Node
     const {
         data: jobs, error: jobsError
-    } = await extractSubgraphv2({
-        'nodeType': 'Company',
-        'name': 'Hirebird',
-    }, sg => sg
-        .extendPath('Company', '<-BELONGS_TO-Job')
-        .extendPath('Company<-BELONGS_TO-Job', '<-SWIPED_ON-User')
-        // .extendPath('Company<-BELONGS_TO-Job<-SWIPED_ON-User', '-BELONGS_TO->Company')
-        // .extendPath('Company<-BELONGS_TO-Job<-SWIPED_ON-User-BELONGS_TO->Company', '<-BELONGS_TO-User')
+    } = await extractSubgraph({
+        'nodeType': 'User',
+        'email': 'root@hirebird.com',
+    }, 
+        sg => sg
+        .extendPath('User', '-BELONGS_TO->Company')
+        .extendPath('User-BELONGS_TO->Company', '<-BELONGS_TO-Job')
+        .extendPath('User-BELONGS_TO->Company<-BELONGS_TO-Job', '<-SWIPED_ON-User')
+        .extendPath('User-BELONGS_TO->Company<-BELONGS_TO-Job<-SWIPED_ON-User', '-BELONGS_TO->Company')
+        .extendPath('User-BELONGS_TO->Company<-BELONGS_TO-Job<-SWIPED_ON-User-BELONGS_TO->Company', '<-BELONGS_TO-User')
+        .extendPath('User', '<-SENT_BY-Message')
     )
     if (jobsError) throwTestError(jobsError)
     // console.log(JSON.stringify(jobs, null, 2))