+++ tests/uix/hb.uix.config.ts
@@ -2,14 +2,15 @@ import { defineConfig } from '@thinairthings/uix'
 import { UserNodeDefinition } from './hb.NodeDefinitions/UserNodeDefinition'
 import { CompanyNodeDefinition } from './hb.NodeDefinitions/CompanyNodeDefinition'
 import { JobNodeDefinition } from './hb.NodeDefinitions/JobNodeDefinition'
-
+import { MessageNodeDefinition } from './hb.NodeDefinitions/MessageNodeDefinition'
 
 export const uixConfig =  defineConfig({
     type: 'Base',
     nodeDefinitionSet: [
         UserNodeDefinition,
         CompanyNodeDefinition,
-        JobNodeDefinition
+        JobNodeDefinition,
+        MessageNodeDefinition
     ],
     outdir: 'tests/uix/generated',
     envPath: '.env.test',